money(double) auto-auditoria: caps de parse de dinero + limpieza float muerto
Auto-auditoria del nucleo (releer diff + grep de simbolos): el compilador NO caza
Float.parseFloat/Float.valueOf/.floatValue() de dinero (devuelven float -> ensanchan
a double pero CAPAN el parseo al techo del float32). Por debajo del techo no se nota
(golden/smoke no lo detectan), pero arruinaria ciegas/cap/apuesta en deep-stack:

  - EditBlindsDialog: parseBlindLevelBB/blindCapSelectedBB/blind_cap -> double;
    actualizarCiegasManualmente con Double.valueOf (era Float.valueOf).
  - GameFrame: BLIND_CAP via Double.parseDouble (era Float.parseFloat).
  - NewGameDialog: CIEGA_GRANDE/PEQUENA via Double.valueOf (era Float.valueOf).
  - LocalPlayer: valor del bet_spinner via doubleValue() (era floatValue()).
  - StatsDialog: sb/sbval/pot via getDouble (era getFloat).

Sin flecos float en el path de dinero: display de buy-in/rebuy a money2String (eran
float2String). Eliminado Helpers.float2String (muerto, cero callers; el dinero usa
money2String). Conservado float1DSecureCompare (lo usa el lado float del bot) +
comentarios de floatClean/doubleClean al dia.

Compila limpio, golden H_final byte-identico, GameFlowSmoke conserva fichas. 118/118.

+++ src/main/java/com/tonikelope/coronapoker/EditBlindsDialog.java
@@ -455,9 +455,9 @@ public class EditBlindsDialog extends JDialog {
 
     // Tope de ciegas como "nº de subidas" (igual que NewGameDialog): el spinner es
     // el entero y blind_cap_label muestra el nivel resultante. GameFrame.BLIND_CAP
-    // sigue siendo la ciega grande de ese nivel (float).
-    private float parseBlindLevelBB(String item) {
-        return Float.parseFloat(item.replace(",", ".").split("/")[1].trim());
+    // sigue siendo la ciega grande de ese nivel (double).
+    private double parseBlindLevelBB(String item) {
+        return Double.parseDouble(item.replace(",", ".").split("/")[1].trim());
     }
 
     private int blindCapTargetIndex(int n) {
@@ -465,7 +465,7 @@ public class EditBlindsDialog extends JDialog {
         return Math.min(Math.max(0, ciegas_combobox.getSelectedIndex()) + n, last);
     }
 
-    private float blindCapSelectedBB() {
+    private double blindCapSelectedBB() {
         return parseBlindLevelBB(ciegas_combobox.getItemAt(blindCapTargetIndex(((Number) blind_cap_spinner.getValue()).intValue())));
     }
 
@@ -513,12 +513,12 @@ public class EditBlindsDialog extends JDialog {
 
         }
 
-        float blind_cap = (this.doblar_checkbox.isSelected() && this.blind_cap_checkbox.isSelected()) ? blindCapSelectedBB() : 0f;
+        double blind_cap = (this.doblar_checkbox.isSelected() && this.blind_cap_checkbox.isSelected()) ? blindCapSelectedBB() : 0;
         GameFrame.BLIND_CAP = blind_cap;
 
         String[] valores_ciegas = ((String) ciegas_combobox.getSelectedItem()).replace(",", ".").split("/");
 
-        GameFrame.getInstance().getCrupier().actualizarCiegasManualmente(Float.valueOf(valores_ciegas[0].trim()), Float.valueOf(valores_ciegas[1].trim()), ciegas_double, ciegas_double_type);
+        GameFrame.getInstance().getCrupier().actualizarCiegasManualmente(Double.valueOf(valores_ciegas[0].trim()), Double.valueOf(valores_ciegas[1].trim()), ciegas_double, ciegas_double_type);
 
         setVisible(false);
